파티클 인스턴싱 추가

+++ DX11_Wizard_of_Legend/ENGINE_SOURCE/jsEntity.cpp
@@ -3,11 +3,11 @@
 namespace js
 {
 	Entity::Entity()
-		: mID(reinterpret_cast<UINT32>(this))
+		: mID((UINT32)reinterpret_cast<UINT32>(this))
 	{
 	}
 	Entity::Entity(const Entity& other)
-		: mID(reinterpret_cast<UINT32>(this))
+		: mID((UINT32)reinterpret_cast<UINT32>(this))
 		, mName(other.mName)
 	{
 	}
